backend/requirements.txt
# PathMind AI - Backend Dependencies
# Install: pip install -r requirements.txt

# Core API
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
python-dotenv>=1.0.1

# Auth / Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
PyJWT>=2.8.0
bcrypt>=4.1.3
slowapi>=0.1.9

# Database
sqlalchemy[asyncio]>=2.0.31
alembic>=1.13.2
asyncpg>=0.29.0
aiosqlite>=0.20.0

# Validation / HTTP
pydantic[email]>=2.7.4
email-validator>=2.2.0
httpx>=0.27.0
pytrends>=4.9.2

# ML / Data
numpy>=1.26.4
pandas>=2.2.2
scikit-learn>=1.5.1
joblib>=1.4.2
spacy>=3.7.5
PyPDF2>=3.0.1

# Scheduling / Payments
apscheduler>=3.10.4
razorpay>=1.4.2

# Tests
pytest>=8.2.0
pytest-asyncio>=0.23.6

# Dev Tools
black>=24.4.2
isort>=5.13.2
mypy>=1.10.0
